firmware/requirements.txt
paho-mqtt==1.6.1
pydantic==1.8.2